+++ backend/openvpnathome/settings/installed_apps.py
@@ -7,6 +7,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'channels',
 
     'rest_framework',
     'rest_framework.authtoken',