+++ bin/bootstrap.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
+
 #
 # This script performs application setup: db and user creation, etc.
 # You should call it after installation to finalize the app deployment.
@@ -54,11 +55,11 @@ bootstrap_default_config() {
 
     CONFIG_ARGS="--accept --admin-email \"${ADMIN}\""
 
-    [ ! -z "${SMTP_SERVER}" ]   && CONFIG_ARGS="${CONFIG_ARGS} --smtp-server \"${SMTP_SERVER}\""
-    [ ! -z "${SMTP_PORT}" ]     && CONFIG_ARGS="${CONFIG_ARGS} --smtp-port \"${SMTP_PORT}\""
-    [ ! -z "${SMTP_LOGIN}" ]    && CONFIG_ARGS="${CONFIG_ARGS} --smtp-login \"${SMTP_LOGIN}\""
-    [ ! -z "${SMTP_PASSWORD}" ] && CONFIG_ARGS="${CONFIG_ARGS} --smtp-password \"${SMTP_PASSWORD}\""
-    [ ! -z "${FORCE}" ]         && CONFIG_ARGS="${CONFIG_ARGS} --force"
+    [[ ! -z "${SMTP_SERVER}"   ]] && CONFIG_ARGS="${CONFIG_ARGS} --smtp-server \"${SMTP_SERVER}\""
+    [[ ! -z "${SMTP_PORT}"     ]] && CONFIG_ARGS="${CONFIG_ARGS} --smtp-port \"${SMTP_PORT}\""
+    [[ ! -z "${SMTP_LOGIN}"    ]] && CONFIG_ARGS="${CONFIG_ARGS} --smtp-login \"${SMTP_LOGIN}\""
+    [[ ! -z "${SMTP_PASSWORD}" ]] && CONFIG_ARGS="${CONFIG_ARGS} --smtp-password \"${SMTP_PASSWORD}\""
+    [[ ! -z "${FORCE}"         ]] && CONFIG_ARGS="${CONFIG_ARGS} --force"
 
     echo "Creating settings.json file."
     eval ${PYTHON} ${MANAGE} configure ${CONFIG_ARGS}
@@ -71,6 +72,9 @@ bootstrap_default_config() {
     eval ${PYTHON} ${MANAGE} set_admin "${ADMIN}" \"${PASSWORD}\"
 }
 
+# Interactively ask for SMTP server configuration using GNU Dialog.
+# It terminates the script if SMTP configuration is not fully
+# provided.
 ask_for_smtp_config() {
     exec 3>&1
     local VALUES=$(dialog --title "SMTP account configuration" \
@@ -80,9 +84,10 @@ ask_for_smtp_config() {
                     "Port (TLS)"  2 1 "${SMTP_PORT}"     2 15 30 0\
                     "Login:"      3 1 "${SMTP_LOGIN}"    3 15 30 0\
                     "Password"    4 1 "${SMTP_PASSWORD}" 4 15 30 0\
-                    2>&1 1>&3)ss
+                    2>&1 1>&3)
     exec 3>&-
 
+    # Values returned in lines - use sed to break it into individual variables
     SMTP_SERVER=$(sed -n 1p <<< ${VALUES})
     SMTP_PORT=$(sed -n 2p <<< ${VALUES})
     SMTP_LOGIN=$(sed -n 3p <<< ${VALUES})
@@ -94,15 +99,18 @@ ask_for_smtp_config() {
     fi
 }
 
+# Check if SMTP configuration is sane.
 is_smtp_configured() {
     # All SMTP options must be set
-    if [ ! "${SMTP_SERVER}" ] || [ ! "${SMTP_PORT}" ] || [ ! "${SMTP_LOGIN}" ] || [ ! "${SMTP_PASSWORD}" ]; then
+    if [[ ! "${SMTP_SERVER}" ]] || [[ ! "${SMTP_PORT}" ]] || [[ ! "${SMTP_LOGIN}" ]] || [[ ! "${SMTP_PASSWORD}" ]]; then
         return 1
     else
         return 0
     fi
 }
 
+# Interactively ask for admin e-mail. It terminates the script
+# if input is empty, but does not validate input otherwise.
 ask_for_admin_email() {
     exec 3>&1
     ADMIN=$(dialog --title "Admin e-mail" \
@@ -110,12 +118,15 @@ ask_for_admin_email() {
                    2>&1 1>&3)
     exec 3>&-
 
-    if [ -z "${ADMIN}" ]; then
+    if [[ -z "${ADMIN}" ]]; then
         echo "Admin e-mail is required"
         exit 1
     fi
 }
 
+# Interactively ask for admin password twice.
+# Passwords must match. If they don't match then
+# terminate the script with error message.
 ask_for_admin_password() {
     exec 3>&1
     PASSWORD=$(dialog --title "Admin account password" \
@@ -129,12 +140,12 @@ ask_for_admin_password() {
                             2>&1 1>&3)
     exec 3>&-
 
-    if [ "${PASSWORD}" != "${PASSWORD_CHECK}" ]; then
+    if [[ "${PASSWORD}" != "${PASSWORD_CHECK}" ]]; then
         echo "Passwords do not match. Plesae try again."
         exit 1
     fi
 
-    if [ ${#PASSWORD} -lt 8 ]; then
+    if [[ ${#PASSWORD} -lt 8 ]]; then
         echo "Password must be at least 8 characters long. Please try again."
         exit 1
     fi
@@ -183,13 +194,13 @@ while true; do
 
 done
 
-if [ -z "$1" ]; then
+if [[ -z "$1" ]]; then
     ask_for_admin_email
 else
     ADMIN="$1"
 fi
 
-if [ -z "$2" ]; then
+if [[ -z "$2" ]]; then
     ask_for_admin_password
 else
     echo "Password supplied from commandline. Make sure it is not stored in shell history."
@@ -199,7 +210,7 @@ fi
 if is_smtp_configured; then
     echo "Using supplied SMTP configuration"
 else
-    if [ "${ASK_FOR_SMTP}" ]; then
+    if [[ "${ASK_FOR_SMTP}" ]]; then
         ask_for_smtp_config
     else
         echo "SMTP configuration skipped"
@@ -209,13 +220,15 @@ fi
 create_database_dir
 create_logs_dir
 bootstrap_default_config
+create_ssh_key
 
-if [ "${USER}" == "root" ]; then
+if [[ "${USER}" == "root" ]]; then
     set_database_permissions
     set_config_permissions
+    set_ssh_permissions
 fi
 
-if [ "${USER}" != "root" ]; then
+if [[ "${USER}" != "root" ]]; then
     echo "Restarting service skipped."
 else
     echo "Restarting server..."