+++ bin/common.sh
@@ -14,6 +14,9 @@ export SETTINGS_FILE="${ROOT_DIR}/settings.json"
 export LOGS_DIR="${ROOT_DIR}/logs"
 export BIN_DIR="${ROOT_DIR}/bin"
 export ANSIBLE_DIR="${ROOT_DIR}/ansible"
+export SSH_KEY_DIR="${ROOT_DIR}/ssh"
+export SSH_DEPLOYMENT_KEY="${SSH_KEY_DIR}/openvpnathome_server_deployment_key"
+export APP_HOME="${ROOT_DIR}/home/"
 
 export PYTHON="${VIRTUALENV_DIR}/bin/python3"
 export MANAGE=${APP_DIR}/manage.py
@@ -36,7 +39,10 @@ check_is_root() {
 create_user() {
     if [ -z $(getent passwd ${APP_USER}) ]; then
         echo "Creating user ${APP_USER}"
-        useradd --system ${APP_USER}
+        useradd --system --home-dir "${APP_HOME}" "${APP_USER}"
+        mkdir -p "${APP_HOME}"
+        chmod 0700 "${APP_HOME}"
+        chown ${OWNERSHIP} "${APP_HOME}"
     fi
 }
 
@@ -63,6 +69,12 @@ set_config_permissions() {
     chmod 0440 "${SETTINGS_FILE}"
 }
 
+set_ssh_permissions() {
+    echo "Setting ssh deployment keys permissions"
+    chown -R ${OWNERSHIP} "${SSH_KEY_DIR}"
+    chmod 0700 "${SSH_KEY_DIR}"
+    chmod 0600 "${SSH_KEY_DIR}"/*
+}
 
 create_logs_dir() {
     echo "Creating logs dir owned by ${OWNERSHIP}"
@@ -80,4 +92,14 @@ remove_user() {
 set_manage_py_interpreter() {
     echo "Setting virtual environment interpreter in ${ROOT_DIR}/backend"
     sed -i 's/\/usr\/bin\/env.*/\/srv\/openvpnathome\/env\/bin\/python3/' "${ROOT_DIR}/backend/manage.py"
+}
+
+create_ssh_key() {
+    mkdir -p "${SSH_KEY_DIR}"
+    if [[ ! -f "${SSH_DEPLOYMENT_KEY}" ]]; then
+        echo "Creating ssh key in ${SSH_KEY_DIR}"
+        ssh-keygen -N '' -C OpenVPN@Home -f "${SSH_DEPLOYMENT_KEY}"
+    else
+        echo "SSH deployment key already created. Skipping."
+    fi
 }